refactor: Replace Translator with Adapter and consolidate API execution logic
