+++ go.mod
@@ -12,7 +12,7 @@ require (
 	github.com/pardnchiu/ToriiDB v0.5.1
 	github.com/pardnchiu/go-bot v0.4.3
 	github.com/pardnchiu/go-browser v0.1.7
-	github.com/pardnchiu/go-pkg v0.13.3
+	github.com/pardnchiu/go-pkg v0.13.4
 	github.com/pardnchiu/go-scheduler v1.2.0
 	github.com/pardnchiu/go-sqlite v1.1.0
 )
