+++ internal/toolAdapter/api/execute.go
@@ -3,46 +3,29 @@ package apiAdapter
 import (
 	"context"
 	"encoding/base64"
-	"encoding/json"
 	"fmt"
-	"io"
 	"log/slog"
-	"net/http"
+	"maps"
+	"net/url"
 	"strings"
-	"sync"
 	"time"
 
 	"github.com/pardnchiu/go-pkg/filesystem/keychain"
+	go_pkg_http "github.com/pardnchiu/go-pkg/http"
 )
 
-const (
-	apiMinGap = 1 * time.Second
-)
-
-var (
-	apiSlotMu   sync.Mutex
-	apiKeyMutex = make(map[string]*sync.Mutex)
-	apiLastCall = make(map[string]time.Time)
-)
-
-func (t *Translator) Execute(ctx context.Context, name string, params map[string]any) (string, error) {
-	key := strings.TrimPrefix(name, t.prefix)
-	doc, ok := t.apis[key]
+func (a *Adapter) Execute(ctx context.Context, name string, params map[string]any) (string, error) {
+	key := strings.TrimPrefix(name, a.prefix)
+	doc, ok := a.apis[key]
 	if !ok {
 		return "", fmt.Errorf("api tool not found: %s", name)
 	}
 
-	normalizeAliasParams(doc, params)
-
-	if err := t.checkRequireds(doc, params); err != nil {
-		return "", fmt.Errorf("t.checkRequireds: %w", err)
-	}
-
-	if err := reserveAPISlot(ctx, key); err != nil {
+	if err := a.checkParams(doc, params); err != nil {
 		return "", err
 	}
 
-	result, err := t.send(ctx, key, doc, params)
+	result, err := a.send(ctx, key, doc, params)
 	if err != nil {
 		return "", fmt.Errorf("t.send: %w", err)
 	}
@@ -50,51 +33,7 @@ func (t *Translator) Execute(ctx context.Context, name string, params map[string
 	return result, nil
 }
 
-func reserveAPISlot(ctx context.Context, key string) error {
-	apiSlotMu.Lock()
-	mu, ok := apiKeyMutex[key]
-	if !ok {
-		mu = &sync.Mutex{}
-		apiKeyMutex[key] = mu
-	}
-	apiSlotMu.Unlock()
-
-	mu.Lock()
-	defer mu.Unlock()
-
-	if wait := apiMinGap - time.Since(apiLastCall[key]); wait > 0 {
-		select {
-		case <-time.After(wait):
-		case <-ctx.Done():
-			return ctx.Err()
-		}
-	}
-	apiLastCall[key] = time.Now()
-	return nil
-}
-
-func normalizeAliasParams(doc *APIDocumentData, params map[string]any) {
-	aliasMap := map[string][]string{
-		"currency": {"base", "from"},
-	}
-
-	for canonical, aliases := range aliasMap {
-		if _, ok := doc.Parameters[canonical]; !ok {
-			continue
-		}
-		if value, ok := params[canonical]; ok && value != nil {
-			continue
-		}
-		for _, alias := range aliases {
-			if value, ok := params[alias]; ok && value != nil {
-				params[canonical] = value
-				break
-			}
-		}
-	}
-}
-
-func (t *Translator) checkRequireds(doc *APIDocumentData, params map[string]any) error {
+func (a *Adapter) checkParams(doc *Document, params map[string]any) error {
 	for name, schema := range doc.Parameters {
 		if _, exists := params[name]; !exists {
 			if schema.Required {
@@ -108,77 +47,127 @@ func (t *Translator) checkRequireds(doc *APIDocumentData, params map[string]any)
 	return nil
 }
 
-func (t *Translator) send(ctx context.Context, key string, doc *APIDocumentData, params map[string]any) (string, error) {
-	var (
-		req *http.Request
-		err error
-	)
+func (a *Adapter) send(ctx context.Context, key string, doc *Document, params map[string]any) (string, error) {
+	apiPath := replaceParams(doc, params)
 
-	switch doc.Endpoint.ContentType {
-	case "form":
-		req, err = t.FormDataRequest(doc, params)
-	default:
-		req, err = t.JSONRequest(doc, params)
-	}
-	if err != nil {
-		return "", fmt.Errorf("build request: %w", err)
-	}
-
-	for k, v := range doc.Endpoint.Headers {
-		req.Header.Set(k, v)
-	}
-
-	// * add auth to header, ex. bearer token, api key, basic auth
-	if doc.Auth != nil && *doc.Auth != (APIDocumentAuthData{}) {
-		if err := t.insetAuth(req, doc.Auth); err != nil {
+	header := make(map[string]string, len(doc.Endpoint.Headers)+1)
+	maps.Copy(header, doc.Endpoint.Headers)
+	if doc.Auth != nil && *doc.Auth != (DocumentAuth{}) {
+		if err := appendAuth(header, doc.Auth); err != nil {
 			return "", err
 		}
 	}
 
-	timeoutSec := doc.Endpoint.Timeout
-	if timeoutSec <= 0 {
-		timeoutSec = 60
+	sec := doc.Endpoint.Timeout
+	if sec <= 0 {
+		sec = 60
 	}
-	timeout := time.Duration(timeoutSec) * time.Second
-
-	reqCtx, cancel := context.WithTimeout(ctx, timeout)
+	reqCtx, cancel := context.WithTimeout(ctx, time.Duration(sec)*time.Second)
 	defer cancel()
-	req = req.WithContext(reqCtx)
 
-	resp, err := t.doWithProgress(reqCtx, t.prefix+key, req, timeoutSec)
-	if err != nil {
-		return "", fmt.Errorf("client.Do: %w", err)
-	}
-	defer resp.Body.Close()
+	name := a.prefix + key
+	resultCh := make(chan struct {
+		body string
+		err  error
+	}, 1)
 
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return "", fmt.Errorf("io.ReadAll: %w", err)
-	}
-
-	if resp.StatusCode != http.StatusOK {
-		return "", fmt.Errorf("resp.StatusCode: %d", resp.StatusCode)
-	}
-
-	if doc.Response.Format == "json" {
-		var data any
-		if err := json.Unmarshal(body, &data); err == nil {
-			output, err := json.Marshal(data)
-			if err != nil {
-				return "", fmt.Errorf("json.Marshal: %w", err)
+	go func() {
+		var (
+			body string
+			err  error
+		)
+		switch doc.Endpoint.Method {
+		case "GET":
+			query := url.Values{}
+			for k, v := range doc.Endpoint.Query {
+				query.Set(k, v)
 			}
-			return string(output), nil
+			for k, v := range params {
+				query.Set(k, fmt.Sprintf("%v", v))
+			}
+			if len(query) > 0 {
+				apiPath = apiPath + "?" + query.Encode()
+			}
+
+			body, _, err = go_pkg_http.GET[string](reqCtx, a.client, apiPath, header)
+		case "POST":
+			body, _, err = go_pkg_http.POST[string](reqCtx, a.client, apiPath, header, params, doc.Endpoint.ContentType)
+		case "PUT":
+			body, _, err = go_pkg_http.PUT[string](reqCtx, a.client, apiPath, header, params, doc.Endpoint.ContentType)
+		case "PATCH":
+			body, _, err = go_pkg_http.PATCH[string](reqCtx, a.client, apiPath, header, params, doc.Endpoint.ContentType)
+		case "DELETE":
+			body, _, err = go_pkg_http.DELETE[string](reqCtx, a.client, apiPath, header, params, doc.Endpoint.ContentType)
+		default:
+			err = fmt.Errorf("unsupported method: %s", doc.Endpoint.Method)
+		}
+		resultCh <- struct {
+			body string
+			err  error
+		}{body, err}
+	}()
+
+	// * every 30s print for checking health
+	ticker := time.NewTicker(30 * time.Second)
+	defer ticker.Stop()
+
+	start := time.Now()
+	for {
+		select {
+		case r := <-resultCh:
+			return r.body, r.err
+
+		case <-ticker.C:
+			slog.Warn("running",
+				slog.String("name", name),
+				slog.String("elapsed", fmt.Sprintf("%ds/%ds", int(time.Since(start).Seconds()), sec)))
+
+		case <-reqCtx.Done():
+			return "", reqCtx.Err()
+		}
+	}
+}
+
+// * https://example.com/post/{id}
+func replaceParams(doc *Document, params map[string]any) string {
+	apiPath := doc.Endpoint.URL
+
+	for k, v := range params {
+		placeholder := "{" + k + "}"
+		if strings.Contains(apiPath, placeholder) {
+			val := fmt.Sprintf("%v", v)
+			if val == "" {
+				delete(params, k)
+				continue
+			}
+
+			apiPath = strings.ReplaceAll(apiPath, placeholder, url.PathEscape(val))
+			delete(params, k)
 		}
 	}
 
-	return string(body), nil
+	for strings.Contains(apiPath, "{") {
+		start := strings.Index(apiPath, "{")
+		end := strings.Index(apiPath, "}")
+		if end < start {
+			break
+		}
+
+		newStart := start
+		if newStart > 0 && apiPath[newStart-1] == '/' {
+			newStart--
+		}
+		apiPath = apiPath[:newStart] + apiPath[end+1:]
+	}
+	return apiPath
 }
 
-func (t *Translator) insetAuth(req *http.Request, auth *APIDocumentAuthData) error {
+func appendAuth(header map[string]string, auth *DocumentAuth) error {
 	if auth.Env == "" {
-		return fmt.Errorf("auth.env is required")
+		return fmt.Errorf("env key is required")
 	}
 
+	// * keychain already contain keychain and env check
 	value := keychain.Get(auth.Env)
 	if value == "" {
 		return fmt.Errorf("%q not set", auth.Env)
@@ -186,52 +175,20 @@ func (t *Translator) insetAuth(req *http.Request, auth *APIDocumentAuthData) err
 
 	switch auth.Type {
 	case "bearer":
-		req.Header.Set("Authorization", "Bearer "+value)
+		header["Authorization"] = "Bearer " + value
 
 	case "apikey":
-		header := auth.Header
-		if header == "" {
-			header = "X-API-Key"
+		key := auth.Header
+		if key == "" {
+			key = "X-API-Key"
 		}
-		req.Header.Set(header, value)
+		header[key] = value
 
 	case "basic":
-		encoded := base64.StdEncoding.EncodeToString([]byte(value))
-		req.Header.Set("Authorization", "Basic "+encoded)
+		header["Authorization"] = "Basic " + base64.StdEncoding.EncodeToString([]byte(value))
 
 	default:
 		return fmt.Errorf("unsupported auth: %s", auth.Type)
 	}
-
 	return nil
-}
-
-type httpResult struct {
-	resp *http.Response
-	err  error
-}
-
-func (t *Translator) doWithProgress(ctx context.Context, name string, req *http.Request, timeoutSec int) (*http.Response, error) {
-	done := make(chan httpResult, 1)
-	go func() {
-		resp, err := t.client.Do(req)
-		done <- httpResult{resp, err}
-	}()
-
-	start := time.Now()
-	ticker := time.NewTicker(30 * time.Second)
-	defer ticker.Stop()
-
-	for {
-		select {
-		case r := <-done:
-			return r.resp, r.err
-		case <-ticker.C:
-			slog.Warn("running",
-				slog.String("name", name),
-				slog.String("elapsed", fmt.Sprintf("%ds/%ds", int(time.Since(start).Seconds()), timeoutSec)))
-		case <-ctx.Done():
-			return nil, ctx.Err()
-		}
-	}
 }
