+++ internal/tools/external/send_test.go
@@ -1,6 +1,4 @@
-// Generated by Claude Code
-
-package apiAdapter
+package external
 
 import (
 	"context"
@@ -127,7 +125,7 @@ func TestBuildMultipart_InvalidFileEntry(t *testing.T) {
 	}
 }
 
-func TestSend_GET(t *testing.T) {
+func TestSendHTTPRequest_GET(t *testing.T) {
 	srv := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		w.Header().Set("Content-Type", "application/json")
 		w.WriteHeader(200)
@@ -135,11 +133,11 @@ func TestSend_GET(t *testing.T) {
 	}))
 	defer srv.Close()
 
-	result, err := Send(context.Background(), srv.URL+"/test", "GET", nil, nil, "json", 5)
+	result, err := sendHTTPRequest(context.Background(), srv.URL+"/test", "GET", nil, nil, "json", 5)
 	if err != nil {
-		t.Fatalf("Send: %v", err)
+		t.Fatalf("sendHTTPRequest: %v", err)
 	}
-	var resp ResponseData
+	var resp httpResponse
 	json.Unmarshal([]byte(result), &resp)
 	if resp.StatusCode != 200 {
 		t.Errorf("status = %d", resp.StatusCode)
@@ -149,7 +147,7 @@ func TestSend_GET(t *testing.T) {
 	}
 }
 
-func TestSend_POST_JSON(t *testing.T) {
+func TestSendHTTPRequest_POST_JSON(t *testing.T) {
 	srv := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		if r.Method != "POST" {
 			t.Errorf("method = %s", r.Method)
@@ -162,13 +160,13 @@ func TestSend_POST_JSON(t *testing.T) {
 	}))
 	defer srv.Close()
 
-	_, err := Send(context.Background(), srv.URL, "POST", nil, map[string]any{"key": "val"}, "json", 5)
+	_, err := sendHTTPRequest(context.Background(), srv.URL, "POST", nil, map[string]any{"key": "val"}, "json", 5)
 	if err != nil {
-		t.Fatalf("Send: %v", err)
+		t.Fatalf("sendHTTPRequest: %v", err)
 	}
 }
 
-func TestSend_CustomHeaders(t *testing.T) {
+func TestSendHTTPRequest_CustomHeaders(t *testing.T) {
 	srv := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		if r.Header.Get("X-Custom") != "value" {
 			t.Errorf("X-Custom = %q", r.Header.Get("X-Custom"))
@@ -178,45 +176,45 @@ func TestSend_CustomHeaders(t *testing.T) {
 	}))
 	defer srv.Close()
 
-	_, err := Send(context.Background(), srv.URL, "GET", map[string]string{"X-Custom": "value"}, nil, "", 5)
+	_, err := sendHTTPRequest(context.Background(), srv.URL, "GET", map[string]string{"X-Custom": "value"}, nil, "", 5)
 	if err != nil {
-		t.Fatalf("Send: %v", err)
+		t.Fatalf("sendHTTPRequest: %v", err)
 	}
 }
 
-func TestSend_InvalidMethod(t *testing.T) {
-	_, err := Send(context.Background(), "https://example.com", "TRACE", nil, nil, "", 5)
+func TestSendHTTPRequest_InvalidMethod(t *testing.T) {
+	_, err := sendHTTPRequest(context.Background(), "https://example.com", "TRACE", nil, nil, "", 5)
 	if err == nil {
 		t.Error("expected error for TRACE method")
 	}
 }
 
-func TestSend_EmptyURL(t *testing.T) {
-	_, err := Send(context.Background(), "", "GET", nil, nil, "", 5)
+func TestSendHTTPRequest_EmptyURL(t *testing.T) {
+	_, err := sendHTTPRequest(context.Background(), "", "GET", nil, nil, "", 5)
 	if err == nil {
 		t.Error("expected error for empty URL")
 	}
 }
 
-func TestSend_TimeoutClamp(t *testing.T) {
+func TestSendHTTPRequest_TimeoutClamp(t *testing.T) {
 	srv := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		w.WriteHeader(200)
 		w.Write([]byte("ok"))
 	}))
 	defer srv.Close()
 
-	_, err := Send(context.Background(), srv.URL, "GET", nil, nil, "", -10)
+	_, err := sendHTTPRequest(context.Background(), srv.URL, "GET", nil, nil, "", -10)
 	if err != nil {
-		t.Fatalf("Send with negative timeout: %v", err)
+		t.Fatalf("sendHTTPRequest with negative timeout: %v", err)
 	}
 
-	_, err = Send(context.Background(), srv.URL, "GET", nil, nil, "", 999)
+	_, err = sendHTTPRequest(context.Background(), srv.URL, "GET", nil, nil, "", 999)
 	if err != nil {
-		t.Fatalf("Send with large timeout: %v", err)
+		t.Fatalf("sendHTTPRequest with large timeout: %v", err)
 	}
 }
 
-func TestSend_POST_Form(t *testing.T) {
+func TestSendHTTPRequest_POST_Form(t *testing.T) {
 	srv := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		if r.Header.Get("Content-Type") != "application/x-www-form-urlencoded" {
 			t.Errorf("content-type = %s", r.Header.Get("Content-Type"))
@@ -226,8 +224,8 @@ func TestSend_POST_Form(t *testing.T) {
 	}))
 	defer srv.Close()
 
-	_, err := Send(context.Background(), srv.URL, "POST", nil, map[string]any{"a": "b"}, "form", 5)
+	_, err := sendHTTPRequest(context.Background(), srv.URL, "POST", nil, map[string]any{"a": "b"}, "form", 5)
 	if err != nil {
-		t.Fatalf("Send form: %v", err)
+		t.Fatalf("sendHTTPRequest form: %v", err)
 	}
 }
